Move RunsSection to ListRuns components
